project-backups/p340-ai-answering/requirements.txt
Hershey_Fonts==2.1.0
numpy==2.3.2
openai==1.98.0
opencv_python==4.11.0.86
Pillow==11.3.0
pymycobot==3.9.9
PyYAML==6.0.2
PyYAML==6.0.2
